+++ server/players/Player.js
@@ -1,7 +1,8 @@
 import { tileToPixel } from '../map/collision.js';
+import { createPlayerStats, statsToJSON } from '../../shared/stats.js';
 
 export class Player {
-  constructor({ id, name, characterClass, x, y }) {
+  constructor({ id, name, characterClass, x, y, stats }) {
     this.id = id;
     this.name = name;
     this.characterClass = characterClass;
@@ -12,6 +13,9 @@ export class Player {
     this.aimX = x;
     this.aimY = y;
     this.moving = false;
+    this.attacking = false;
+
+    Object.assign(this, stats);
   }
 
   toJSON() {
@@ -26,11 +30,14 @@ export class Player {
       aimX: this.aimX,
       aimY: this.aimY,
       moving: this.moving,
+      attacking: this.attacking,
+      ...statsToJSON(this),
     };
   }
 }
 
 export function createPlayer({ id, name, characterClass, spawn }) {
   const { x, y } = tileToPixel(spawn.x, spawn.y);
-  return new Player({ id, name, characterClass, x, y });
+  const stats = createPlayerStats(characterClass);
+  return new Player({ id, name, characterClass, x, y, stats });
 }